adobework-backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# PDF Processing - Advanced
PyPDF2>=3.0.0
pdf2docx>=0.5.0
pikepdf>=8.0.0
reportlab>=4.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=0.6.0
pdf2image>=1.16.0
pymupdf>=1.23.0

# Image Processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# OCR
pytesseract>=0.3.10

# Utilities
aiofiles>=23.0.0
